updated temporal alignment
