+++ moore_2025/convert_session.py
@@ -1,16 +1,12 @@
 """Primary script to convert each example session to NWB."""
 from pathlib import Path
-import pandas as pd
 import shutil
 from neuroconv.utils.dict import load_dict_from_file, dict_deep_update
 import woodcode.nwb as nwb
 import probeinterface as pbi
 import numpy as np
-import re
-from datetime import datetime
-import pytz
-from spikeinterface.extractors import OpenEphysBinaryRecordingExtractor
-from scipy.signal import correlate, correlation_lags
+
+from moore_2025.temporal_alignment import get_aligned_video_timestamps_juveniles, get_aligned_video_timestamps_adults, get_start_time
 
 def get_probe_info_juveniles() -> dict:
     # TODO: add logic for daily probe advancement after multi-day example data gets shared. 
@@ -81,224 +77,6 @@ def get_probe_info_adults() -> dict:
         },
     }
     return probe_info
-
-def get_start_time(timestamps_file_path: Path) -> str:
-    """
-    Get the session start datetime from the name of the timestamps CSV file.
-    
-    Parameters
-    ----------
-    timestamps_file_path : Path
-        Path to the timestamps CSV file.
-    
-    Returns
-    -------
-    datetime
-        The session start datetime with timezone information.
-    """
-    # Example filename: 'Bonsai testing2021-08-05T17_06_23.csv'
-    filename = timestamps_file_path.stem  # Ex. 'Bonsai testing2021-08-05T17_06_23'
-    pattern = r'(\d{4}-\d{2}-\d{2}T\d{2}_\d{2}_\d{2})'
-    match = re.search(pattern, filename)
-    if not match:
-        raise ValueError(f"Could not extract datetime from filename: {filename}")
-    
-    start_time = match.group(1)
-    start_time = datetime.strptime(start_time, '%Y-%m-%dT%H_%M_%S')
-    tz_info = pytz.timezone('Europe/London')
-    start_time = tz_info.localize(start_time)
-
-    return start_time
-
-def get_ttl_timestaps(*, traces: np.ndarray, timestamps: np.ndarray, threshold: float, cooldown_in_seconds: float, sampling_rate: float) -> np.ndarray:
-    """
-    Get TTL timestamps from traces.
-
-    Parameters
-    ----------
-    traces : np.ndarray
-        The input signal traces.
-    timestamps : np.ndarray
-        The corresponding timestamps for the traces.
-    threshold : float
-        The threshold to detect TTL onsets and offsets.
-    cooldown_in_seconds : float
-        The cooldown period in seconds to avoid detecting the same onset multiple times.
-    sampling_rate : float
-        The sampling rate of the traces in Hz.
-    
-    Returns
-    -------
-    np.ndarray
-        The timestamps of detected TTL centers.
-    """    
-
-    peak_points = np.where((traces > threshold))[0]
-    gaps = np.diff(peak_points) > 1
-    onsets = np.concatenate([[peak_points[0]], peak_points[1:][gaps]])
-    offsets = np.concatenate([peak_points[:-1][gaps], [peak_points[-1]]])
-    onsets = np.array(onsets)
-    offsets = np.array(offsets)
-    centers = (onsets + offsets) // 2
-
-    # Add a cooldown period to avoid detecting the same onset multiple times
-    cooldown_in_samples = int(cooldown_in_seconds * sampling_rate)
-    center_diff = np.diff(centers)
-    centers = np.concatenate([[centers[0]], centers[1:][center_diff > cooldown_in_samples]])
-
-    ttl_timestamps = timestamps[centers]
-    return ttl_timestamps
-
-def align_by_interpolation(unaligned_dense_timestamps: np.ndarray, unaligned_sparse_timestamps: np.ndarray, aligned_sparse_timestamps: np.ndarray) -> np.ndarray:
-    """
-    Interpolate timestamps using a mapping from an unaligned time basis to an aligned one.
-
-    Use this function when the unaligned timestamps of data are not directly tracked by a primary
-    system, but are known to occur between timestamps that are tracked. The timestamps are aligned
-    by interpolating between the two time bases.
-
-    An example could be a metronomic TTL pulse (e.g., every second) from a secondary data stream to the primary
-    timing system; if the time references are recorded within the relative time of the secondary
-    data stream, then their exact time in the primary system is inferred given the pulse times.
-
-    Parameters
-    ----------
-    unaligned_dense_timestamps : np.ndarray
-        The dense timestamps of the unaligned secondary time basis that need to be aligned.
-    unaligned_sparse_timestamps : np.ndarray
-        The timestamps of the unaligned secondary time basis.
-    aligned_sparse_timestamps : np.ndarray
-        The timestamps aligned to the primary time basis.
-
-    Returns
-    -------
-    np.ndarray
-        The aligned dense timestamps.
-    """
-    aligned_dense_timestamps = np.interp(
-        x=unaligned_dense_timestamps,
-        xp=unaligned_sparse_timestamps,
-        fp=aligned_sparse_timestamps
-    )
-    return aligned_dense_timestamps
-
-def get_aligned_video_timestamps_juveniles(
-    *,
-    timestamp_file_path: Path,
-    ephys_folder_path: Path,
-) -> np.ndarray:
-    """
-    Get aligned video timestamps for juvenile sessions.
-
-    Parameters
-    ----------
-    timestamp_file_path : Path
-        Path to the video timestamps CSV file.
-    ephys_folder_path : Path
-        Path to the ephys OpenEphys record node folder.
-
-    Returns
-    -------
-    np.ndarray
-        The aligned video timestamps.
-    """
-    led_threshold = 2_000
-    video_sampling_rate = 30.0
-    timestamp_column_name = "Item4.Timestamp"
-    led_column_name = "Item3.Val0"
-    ttl_threshold = 20_000
-    ttl_stream_name = "Rhythm_FPGA-100.0_ADC"
-    ttl_channel_id = 'ADC1'
-    cooldown_in_seconds = 0.5
-
-    timestamps_df = pd.read_csv(timestamp_file_path, parse_dates=[timestamp_column_name])
-    traces = timestamps_df[led_column_name].values
-    video_timestamps = np.arange(traces.shape[0]) / video_sampling_rate
-    led_timestamps = get_ttl_timestaps(traces=traces, timestamps=video_timestamps, threshold=led_threshold, cooldown_in_seconds=cooldown_in_seconds, sampling_rate=video_sampling_rate)
-    led_intervals = np.diff(led_timestamps)
-
-    extractor = OpenEphysBinaryRecordingExtractor(folder_path=ephys_folder_path, stream_name=ttl_stream_name)
-    sampling_rate = extractor.get_sampling_frequency()
-    ttl_timestamps = np.ones_like(led_timestamps) * np.nan
-    for segment_index in range(extractor.get_num_segments()):
-        traces = extractor.get_traces(segment_index=segment_index, channel_ids=[ttl_channel_id])
-        ephys_timestamps = extractor.get_times(segment_index=segment_index)
-        single_segment_ttl_timestamps = get_ttl_timestaps(traces=traces, timestamps=ephys_timestamps, threshold=ttl_threshold, cooldown_in_seconds=cooldown_in_seconds, sampling_rate=sampling_rate)
-        ttl_intervals = np.diff(single_segment_ttl_timestamps)
-        correlation = correlate(led_intervals, ttl_intervals, mode='full')
-        lags = correlation_lags(len(led_intervals), len(ttl_intervals), mode='full')
-        best_lag = lags[np.argmax(correlation)]
-        assert np.all(np.isnan(ttl_timestamps[best_lag:best_lag + len(single_segment_ttl_timestamps)])), f"Overlap in TTL timestamps at segment {segment_index}"
-        ttl_timestamps[best_lag:best_lag + len(single_segment_ttl_timestamps)] = single_segment_ttl_timestamps
-        error = led_intervals[best_lag:best_lag + len(ttl_intervals)] - ttl_intervals
-        assert np.max(np.abs(error)) < 1.0, f"Alignment error too large: {np.max(np.abs(error))} seconds for segment {segment_index}"
-
-    # NaN values represent LED pulses that were not recorded in the ephys data (ex. between segments)
-    not_nan = ~np.isnan(ttl_timestamps)
-    led_timestamps = led_timestamps[not_nan]
-    ttl_timestamps = ttl_timestamps[not_nan]
-
-    aligned_video_timestamps = align_by_interpolation(
-        unaligned_dense_timestamps=video_timestamps,
-        unaligned_sparse_timestamps=led_timestamps,
-        aligned_sparse_timestamps=ttl_timestamps,
-    )
-
-    return aligned_video_timestamps
-
-
-def get_aligned_video_timestamps_adults(
-    *,
-    timestamp_file_paths: list[Path],
-    ephys_folder_path: Path,
-) -> list[np.ndarray]:
-    """
-    Get aligned video timestamps for adult sessions.
-
-    Parameters
-    ----------
-    timestamp_file_paths : list[Path]
-        List of paths to the video timestamps CSV files.
-    ephys_folder_path : Path
-        Path to the ephys OpenEphys record node folder.
-
-    Returns
-    -------
-    list[np.ndarray]
-        The aligned video timestamps for each segment/file.
-    """
-    timestamp_column_name = "Item3.Timestamp"
-    ttl_threshold = 10_000
-    ttl_stream_name = "Rhythm_FPGA-103.0_ADC"
-    ttl_channel_id = 'ADC6'
-    cooldown_in_seconds = 0.0
-
-    extractor = OpenEphysBinaryRecordingExtractor(folder_path=ephys_folder_path, stream_name=ttl_stream_name)
-    sampling_rate = extractor.get_sampling_frequency()
-    assert extractor.get_num_segments() == len(timestamp_file_paths), f"Number of ephys segments ({extractor.get_num_segments()}) does not match number of timestamp files ({len(timestamp_file_paths)})."
-    all_aligned_video_timestamps = []
-
-    for segment_index, timestamp_file_path in enumerate(timestamp_file_paths):
-        timestamps_df = pd.read_csv(timestamp_file_path, parse_dates=[timestamp_column_name])
-        num_frames = timestamps_df.shape[0] + 1 # + 1 bc video has one extra frame at the end
-        traces = extractor.get_traces(segment_index=segment_index, channel_ids=[ttl_channel_id])
-        ephys_timestamps = extractor.get_times(segment_index=segment_index)
-        single_segment_ttl_timestamps = get_ttl_timestaps(traces=traces, timestamps=ephys_timestamps, threshold=ttl_threshold, cooldown_in_seconds=cooldown_in_seconds, sampling_rate=sampling_rate)
-        num_ttls = single_segment_ttl_timestamps.shape[0]
-        assert num_ttls >= num_frames, f"Number of TTLs ({num_ttls}) is less than number of video frames ({num_frames}) for segment {segment_index}."
-
-        # Correct for dropped frames by removing the TTLs that correspond to the largest gaps
-        num_dropped_frames = num_ttls - num_frames
-        timestamps_df["seconds"] = (timestamps_df[timestamp_column_name] - timestamps_df[timestamp_column_name].iloc[0]).dt.total_seconds()
-        t_diff = np.diff(timestamps_df["seconds"])
-        drop_indices = np.argsort(t_diff)[-num_dropped_frames:]  # indices of largest gaps
-        drop_indices = np.sort(drop_indices)  # sort chronologically
-        ttl_indices_to_remove = drop_indices + 1 # since drop happens AFTER that CSV frame
-        single_segment_ttl_timestamps = np.delete(single_segment_ttl_timestamps, ttl_indices_to_remove)
-
-        all_aligned_video_timestamps.append(single_segment_ttl_timestamps)
-
-    return all_aligned_video_timestamps
     
 
 def session_to_nwb(
