+++ pyproject.toml
@@ -51,4 +51,4 @@ Homepage = "https://github.com/Wood-Dudchenko-lab/woodcode"
 "Bug Tracker" = "https://github.com/Wood-Dudchenko-lab/woodcode/issues"
 
 [tool.setuptools.packages.find]
-include = ["woodcode", "woodcode.*"]
+include = ["woodcode", "woodcode.*", "moore_2025"]
